Fixed `publish` command to respect `--no-interaction` option as reported in issue #69
When there is an error with any of the arguments passed to the `publish` command
(the book doesn't exist, the edition doesn't exist, etc.) **easybook** tries to
recover from the error and it allows the user to type in another book/edition
(up to 5 attempts). This behavior improves usability, but it must also comply
with the `-no-interaction` command.

After the changes introduced in this commit, this is the new behavior of the `publish` command:

---

**Command:**

`./book publish inexistent-book inexistent-edition`

**Result:**

```
 ERROR  The given 'inexistent-book' slug doesn't match any book in
 /Users/javier/sf/easybook/doc/ directory

 Please, type the slug of the book (e.g. the-origin-of-species)
 >

 (up to 5 attempts)

```

---

**Command:**

`./book publish inexistent-book inexistent-edition --no-interaction`

**Result:**

```
[RuntimeException]
  ERROR: The directory of the book cannot be found.
   Check that '/Users/javier/sf/easybook/doc' directory
   has a folder named as the book slug ('inexistent-book')
```

---

**Command:**

`./book publish this-book-exists inexistent-edition`

**Result:**

```
 ERROR  The 'inexistent-edition' edition isn't defined for 'this-book-exists' book

 Check that /Users/javier/sf/easybook/doc/this-book-exists/config.yml file
 defines a 'inexistent-edition' edition under the editions option.

 Please, type the name of the edition to be published:
 >

 (up to 5 attempts)
```

---

**Command:**

`./book publish this-book-exists inexistent-edition --no-interaction`

**Result:**

```
[RuntimeException]
  ERROR: The 'inexistent-edition' edition isn't defined for
  'this-book-exists' book.
```