[WIP] Added a new 'chunk_level' option to 'html_chunked' format.
This option allows to improve the publication of complex books as
websites. The default value (`1`) is backwards-compatible and it
just publishes a single HTML page for each <h1> section.

The new `2` value allows to split each chapter/appendix into
several HTML pages, one for each <h1> and <h2> sections.

This is a work-in-progress feature. Therefore, the current code
is too complex (although it works), it's not documented (yet)
and it lacks (more) unit tests. To try it, just add the `chunk_level`
option in any edition:

book:
    # ...
    editions:
        my_edition:
            format:      html_chunked
            chunk_level: 2  # only 1 and 2 are allowed