[WIP] Tweaked the chunk process for books published as websites
This commit removes all the logic from the Twig templates and greatly
simplifies the generation of TOC and navigation URL. All the navigation
URLs are now generated as absolute instead of relative.

I still consider this feature as WIP because I need to try it with
more books and it needs to be properly documented and tested.