Refactored the Slugger class.
This refactor originated when I tried to create a unit test for the
Slugger class. The test required weird dependencies, such as the whole
Silex\Application object and it was very complicated to test some features
(such as generating unique slugs).

The new Slugger class is completely decoupled from the Silex\Application object
and it only tries to do one single thing: generate slugs. Advanced features,
such as ensuring the uniqueness of the slugs, has been pushed to the Application
itself.

In addition, the Slugger now uses the new PHP 5.4 transliterator when available
(this built-in transliterator improves both the quality and the performance
of slug generation, specially for non-latin characters).