Fixed, once and for all, the handling of internal links.
Internal links are the links that point to a specific book chapter or
section. Given that easybook publishes books as websites, PDF files and
EPUB ebooks, internal link resolution is a non-trivial task.

Originally we tried to solve this problem by resolving each internal
link while generating the book itself. The problem is that some internal
links cannot be resolved until the book is completely generated. The
reason is that easybook merges first-level empty section with their
following second-level section (see HtmlChunkedPublisher code: if a
chapter contains no content between the <h1> title and the first <h2>
section, this <h2> section is merged with the previous <h1> section
to generate a better looking book).

Now we solve this problem in a much more simple way. First the book is
completely generated with the wrong internal links. Then, the new
fixInternalLinks() method parses every HTML file generated, looks
for all the internal link and modifies them all to point to the
correct chapter or section.

Besides being a simple and fast solution, this new method allows to
remove the previous LinkPlugin, adding a nice 3% performance improvement.