Added a new benchmark script to improve easybook performance
To make the benchmark as realistic as possible, a real book has
been added to the app/Resources/Books/ directory.

The benchmark publishes several times in a row each edition
of the book and displays the results in a easy to understand
score normalized in a 0 to 100 scale.

The benchmark uses the cool new ProgressHelper and TableHelper from
the Symfony Console Component to display its results.

The following are the scores I got for the benchmark:

+---------+--------------+----------------+-------+
| Format  | Time (msec.) | Memory (bytes) | Score |
+---------+--------------+----------------+-------+
| ebook   | 1,712.89     | 1,572,864.00   | 44.32 |
| web     | 1,583.41     | 1,572,864.00   | 47.78 |
| website | 1,717.64     | 1,572,864.00   | 44.20 |
| kindle  | 3,567.66     | 1,572,864.00   | 47.43 |
| print   | 9,892.98     | 1,572,864.00   | 57.09 |
+---------+--------------+----------------+-------+

 YOUR SCORE: 48.16 / 100.00
             ~~~~~~~~~~~~~~