Improved the publication of books as websites
Until now, books published as websites only included their chapters and
appendices. Other special elements such as the license, the author
or the list of figures were displayed in a special way or they were
simply discarded.

This was completely wrong, because it was based on personal opinions about
how a book should be displayed. As easybook is a generic tool to publish
any kind of book, we should never assume how a book should be displayed.
For that reason, books published as websites now include every book item
and the table of contents also displays them all.

If you don't want some element to appear on your website index, tweak the
`index.twig` template to remove it.