Removed the validateConfiguration() method.
It's "impossible" to validate the book configuration because
the author can really set anything in each of the book fields
and he/she can even define any number of new fields.

In the future, it coud be possible to restrict some of the most
important fields, such as the title, the author/s name/s and the
publication date.