Refactored the code highlighting feature of easybook
This refactorization is backwards compatible and you don't
have to make any change to keep highlighting the code of your
books. The difference is that the internal code now is much better.
Most of the code has been removed from the Application class and has
been moved to the CodeHighlighterServiceProvider and the CodeHighlighter
classes.