Replaced the Zip PHP extension by the OS zip command
The ePub format requires that the 'mimetype' file is
at the first file of the ZIP file and it's uncompressed.
The PHP Zip extension doesn't support compression levels,
so you cannot generate a valid ePub book with PHP.

The method now tries to use the OS 'zip' command if
available and then fallbacks to 'Zip' extension.

(this error prevented the ePub from being validated)
(see #143)