Fixed an error with headers for books published as 1-page-per-chapter
When the chunk_level of a book published as a website is 1, each chapter
is published in one page. <h1> and <h2> headers were correctly published
but <h3> .. <h6> headers weren't properly published because they lacked
the `id` attribute.

Now all the headers are correctly generated regardless of their level.